Move service to shared
