+++ spec/services/group_services/iterates_assignment_groups_to_create_grades_spec.rb
@@ -1,5 +1,5 @@
 require "light-service"
-require "./app/services/group_services/iterates_assignment_groups_to_create_grades"
+require "./app/services/shared/iterates_assignment_groups_to_create_grades"
 
 describe Services::Actions::IteratesAssignmentGroupsToCreateGrades do
   let(:assignment) { create(:assignment) }
